GoToRegionFacts should be a separate file
